requirements.txt
# ML
pandas
numpy
scikit-learn
seaborn
matplotlib
# GenAI
python-dotenv
gradio
openai
openai-agents